requirements.txt
asgiref==3.4.1
astroid==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
Django==4.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysql-connector-python==8.0.27
Pillow==8.4.0
platformdirs==2.4.0
pylint==2.12.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
